Trait pallet_traits::AMM

pub trait AMM<AccountId, CurrencyId, Balance> {
    fn get_amounts_out(
        amount_in: Balance, 
        path: Vec<CurrencyId>
    ) -> Result<Vec<Balance>, DispatchError>;

    fn get_amounts_in(
        amount_out: Balance, 
        path: Vec<CurrencyId>
    ) -> Result<Vec<Balance>, DispatchError>;

    fn swap(
        who: &AccountId, 
        pair: (CurrencyId, CurrencyId), 
        amount_in: Balance
    ) -> Result<(), DispatchError>;

    fn get_pools() -> Result<Vec<(CurrencyId, CurrencyId)>, DispatchError>;
}

Exported traits from our AMM pallet. These functions are to be used by the router to enable multi route token swaps

Required methods

fn get_amounts_out(
amount_in: Balance,
path: Vec<CurrencyId>
) -> Result<Vec<Balance>, DispatchError>

Based on the path specified and the available pool balances this will return the amounts outs when trading the specified amount in

fn get_amounts_in(
amount_out: Balance,
path: Vec<CurrencyId>
) -> Result<Vec<Balance>, DispatchError>

Based on the path specified and the available pool balances this will return the amounts in needed to produce the specified amount out

fn swap(
who: &AccountId,
pair: (CurrencyId, CurrencyId),
amount_in: Balance
) -> Result<(), DispatchError>

Handles a “swap” on the AMM side for “who”. This will move the amount_in funds to the AMM PalletId, trade pair.0 to pair.1 and return a result with the amount of currency that was sent back to the user.

fn get_pools() -> Result<Vec<(CurrencyId, CurrencyId)>, DispatchError>